requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.1
django-allauth==65.1.0
django-extensions==3.2.3
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.0.0
pycparser==2.22
pydot==3.0.2
PyJWT==2.9.0
pyparsing==3.2.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
matplotlib==3.8.2

# PostgreSQL adapter for Django
psycopg2-binary>=2.9,<3.0
